ssd: add window drop-shadows
Add optional drop-shadows to windows using server-side decoration.
The presence and appearance of the shadows can be configured in rc.xml
separately for active and inactive windows.  The default is no shadows
to preserve current behaviour.

The shadows are drawn in two fixed buffers containing a corner and edge,
then the edges are scaled according to window size.